+++ qmat/playgrounds/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Folders containing different experiments performed with `qmat`.
+
+    📣 Codes in those folder are not tested by the CI pipeline.
+"""
