+++ qmat/solvers/__init__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Solvers implementations that can make use of `qmat`-generated coefficients.
+"""
